split dashboard_helper into 3 modules + orig
